+++ src/awesome-tools/replaceNums.js
@@ -1,0 +1,3 @@
+const hideNumbers = (str) => str.replace(/[0-9](?=\d{4})/g, "*");
+
+export default hideNumbers
