+++ src/components/Rating.js
@@ -3,22 +3,13 @@ import StarIcon from '@material-ui/icons/Star';
 import StarBorderIcon from '@material-ui/icons/StarBorder';
 
 export default function Rating({value}) {
-
-    const rate = Math.round(Number(value));
-
-    let stars = []
-    let count = 0;
-
-    while (count < 5) {
-        if (count < rate) {
-            stars.push(<StarIcon key={stars}/>)
-            count ++;
-        } else {
-            stars.push(<StarBorderIcon key={stars}/>)
-            count ++;
-        }
-    }
-
+        const rate = Math.round(Number(value));
+        
+        const stars = [...Array(5)].map((elem, idx) => 
+            idx < rate ? 
+            <StarIcon key={idx}/> :
+            <StarBorderIcon key={idx}/>
+        )
 
     return (
         <div>
